+++ dapr/actor/__init__.py
@@ -6,5 +6,5 @@ Licensed under the MIT License.
 """
 
 from dapr.actor.actor_interface import ActorInterface
-from dapr.actor.client.proxy import ActorProxy, ActorProxyFactory
+from dapr.actor.client.proxy import ActorProxy, DefaultActorProxyFactory
 from dapr.actor.id import ActorId
